moved old device id parsing
