+++ src/modules/fancyzones/FancyZonesLib/JsonHelpers.cpp
@@ -74,6 +74,169 @@ namespace NonLocalizable
     const wchar_t Monitors[] = L"monitors";
 }
 
+namespace BackwardsCompatibility
+{
+    std::optional<DeviceIdData> DeviceIdData::ParseDeviceId(const std::wstring& str)
+    {
+        DeviceIdData data;
+
+        std::wstring temp;
+        std::wstringstream wss(str);
+
+        /*
+        Important fix for device info that contains a '_' in the name:
+        1. first search for '#'
+        2. Then split the remaining string by '_'
+        */
+
+        // Step 1: parse the name until the #, then to the '_'
+        if (str.find(L'#') != std::string::npos)
+        {
+            std::getline(wss, temp, L'#');
+
+            data.deviceName = temp;
+
+            if (!std::getline(wss, temp, L'_'))
+            {
+                return std::nullopt;
+            }
+
+            data.deviceName += L"#" + temp;
+        }
+        else if (std::getline(wss, temp, L'_') && !temp.empty())
+        {
+            data.deviceName = temp;
+        }
+        else
+        {
+            return std::nullopt;
+        }
+
+        // Step 2: parse the rest of the id
+        std::vector<std::wstring> parts;
+        while (std::getline(wss, temp, L'_'))
+        {
+            parts.push_back(temp);
+        }
+
+        if (parts.size() != 3 && parts.size() != 4)
+        {
+            return std::nullopt;
+        }
+
+        /*
+        Refer to FancyZonesUtils::GenerateUniqueId parts contain:
+        1. monitor id [string]
+        2. width of device [int]
+        3. height of device [int]
+        4. virtual desktop id (GUID) [string]
+        */
+        try
+        {
+            for (const auto& c : parts[0])
+            {
+                std::ignore = std::stoi(std::wstring(&c));
+            }
+
+            for (const auto& c : parts[1])
+            {
+                std::ignore = std::stoi(std::wstring(&c));
+            }
+
+            data.width = std::stoi(parts[0]);
+            data.height = std::stoi(parts[1]);
+        }
+        catch (const std::exception&)
+        {
+            return std::nullopt;
+        }
+
+        if (!SUCCEEDED(CLSIDFromString(parts[2].c_str(), &data.virtualDesktopId)))
+        {
+            return std::nullopt;
+        }
+
+        if (parts.size() == 4)
+        {
+            data.monitorId = parts[3]; //could be empty
+        }
+
+        return data;
+    }
+    
+    bool DeviceIdData::IsValidDeviceId(const std::wstring& str)
+    {
+        std::wstring monitorName;
+        std::wstring temp;
+        std::vector<std::wstring> parts;
+        std::wstringstream wss(str);
+
+        /*
+        Important fix for device info that contains a '_' in the name:
+        1. first search for '#'
+        2. Then split the remaining string by '_'
+        */
+
+        // Step 1: parse the name until the #, then to the '_'
+        if (str.find(L'#') != std::string::npos)
+        {
+            std::getline(wss, temp, L'#');
+
+            monitorName = temp;
+
+            if (!std::getline(wss, temp, L'_'))
+            {
+                return false;
+            }
+
+            monitorName += L"#" + temp;
+            parts.push_back(monitorName);
+        }
+
+        // Step 2: parse the rest of the id
+        while (std::getline(wss, temp, L'_'))
+        {
+            parts.push_back(temp);
+        }
+
+        if (parts.size() != 4)
+        {
+            return false;
+        }
+
+        /*
+        Refer to FancyZonesUtils::GenerateUniqueId parts contain:
+        1. monitor id [string]
+        2. width of device [int]
+        3. height of device [int]
+        4. virtual desktop id (GUID) [string]
+        */
+        try
+        {
+            //check if resolution contain only digits
+            for (const auto& c : parts[1])
+            {
+                std::ignore = std::stoi(std::wstring(&c));
+            }
+            for (const auto& c : parts[2])
+            {
+                std::ignore = std::stoi(std::wstring(&c));
+            }
+        }
+        catch (const std::exception&)
+        {
+            return false;
+        }
+
+        if (!FancyZonesUtils::IsValidGuid(parts[3]) || parts[0].empty())
+        {
+            return false;
+        }
+
+        return true;
+    }
+}
+    
 namespace
 {
     json::JsonArray NumVecToJsonArray(const std::vector<int>& vec)
@@ -115,13 +278,16 @@ namespace
         }
 
         std::wstring deviceIdStr = json.GetNamedString(NonLocalizable::DeviceIdStr).c_str();
-        auto deviceId = FancyZonesDataTypes::DeviceIdData::ParseDeviceId(deviceIdStr);
+        auto deviceId = BackwardsCompatibility::DeviceIdData::ParseDeviceId(deviceIdStr);
         if (!deviceId.has_value())
         {
             return std::nullopt;
         }
 
-        data.deviceId = *deviceId;
+        data.deviceId = FancyZonesDataTypes::DeviceIdData{
+            .deviceName = deviceId->deviceName,
+            .virtualDesktopId = deviceId->virtualDesktopId
+        };
         data.zoneSetUuid = json.GetNamedString(NonLocalizable::ZoneSetUuidStr);
 
         if (!FancyZonesUtils::IsValidGuid(data.zoneSetUuid))
@@ -379,7 +545,7 @@ namespace JSONHelpers
             DeviceInfoJSON result;
 
             std::wstring deviceIdStr = device.GetNamedString(NonLocalizable::DeviceIdStr).c_str();
-            auto deviceId = FancyZonesDataTypes::DeviceIdData::ParseDeviceId(deviceIdStr);
+            auto deviceId = BackwardsCompatibility::DeviceIdData::ParseDeviceId(deviceIdStr);
             if (!deviceId.has_value())
             {
                 return std::nullopt;
@@ -539,7 +705,16 @@ namespace JSONHelpers
             layout.SetNamedValue(NonLocalizable::AppliedLayoutsIds::SensitivityRadiusID, json::value(data.sensitivityRadius));
 
             json::JsonObject obj{};
-            obj.SetNamedValue(NonLocalizable::AppliedLayoutsIds::DeviceIdID, json::value(deviceID.toString()));
+            json::JsonObject device{};
+            device.SetNamedValue(NonLocalizable::AppliedLayoutsIds::MonitorID, json::value(deviceID.deviceName));
+
+            auto virtualDesktopStr = FancyZonesUtils::GuidToString(deviceID.virtualDesktopId);
+            if (virtualDesktopStr)
+            {
+                device.SetNamedValue(NonLocalizable::AppliedLayoutsIds::VirtualDesktopID, json::value(virtualDesktopStr.value()));
+            }
+
+            obj.SetNamedValue(NonLocalizable::AppliedLayoutsIds::DeviceID, device);
             obj.SetNamedValue(NonLocalizable::AppliedLayoutsIds::AppliedLayoutID, layout);
 
             layoutsArray.Append(obj);
