+++ src/modules/fancyzones/FancyZonesLib/JsonHelpers.h
@@ -8,6 +8,36 @@
 #include <vector>
 #include <unordered_map>
 
+namespace BackwardsCompatibility
+{
+    struct DeviceIdData
+    {
+        std::wstring deviceName = L"FallbackDevice";
+        int width;
+        int height;
+        GUID virtualDesktopId;
+        std::wstring monitorId;
+
+        static std::optional<DeviceIdData> ParseDeviceId(const std::wstring& str);
+        static bool IsValidDeviceId(const std::wstring& str);
+    };
+
+    inline bool operator==(const BackwardsCompatibility::DeviceIdData& lhs, const BackwardsCompatibility::DeviceIdData& rhs)
+    {
+        return lhs.deviceName.compare(rhs.deviceName) == 0 && lhs.width == rhs.width && lhs.height == rhs.height && (lhs.virtualDesktopId == rhs.virtualDesktopId || lhs.virtualDesktopId == GUID_NULL || rhs.virtualDesktopId == GUID_NULL) && lhs.monitorId.compare(rhs.monitorId) == 0;
+    }
+
+    inline bool operator!=(const BackwardsCompatibility::DeviceIdData& lhs, const BackwardsCompatibility::DeviceIdData& rhs)
+    {
+        return !(lhs == rhs);
+    }
+
+    inline bool operator<(const BackwardsCompatibility::DeviceIdData& lhs, const BackwardsCompatibility::DeviceIdData& rhs)
+    {
+        return lhs.deviceName.compare(rhs.deviceName) < 0 || lhs.width < rhs.width || lhs.height < rhs.height || lhs.monitorId.compare(rhs.monitorId) < 0;
+    }
+}
+
 namespace JSONHelpers
 {
     namespace CanvasLayoutInfoJSON
@@ -39,10 +69,9 @@ namespace JSONHelpers
 
     struct DeviceInfoJSON
     {
-        FancyZonesDataTypes::DeviceIdData deviceId;
+        BackwardsCompatibility::DeviceIdData deviceId;
         FancyZonesDataTypes::DeviceInfoData data;
 
-        static json::JsonObject ToJson(const DeviceInfoJSON& device);
         static std::optional<DeviceInfoJSON> FromJson(const json::JsonObject& device);
     };
 
@@ -55,8 +84,7 @@ namespace JSONHelpers
         static std::optional<LayoutQuickKeyJSON> FromJson(const json::JsonObject& device);
     };
 
-    using TAppZoneHistoryMap = std::unordered_map<std::wstring, std::vector<FancyZonesDataTypes::AppZoneHistoryData>>;
-    using TDeviceInfoMap = std::unordered_map<FancyZonesDataTypes::DeviceIdData, FancyZonesDataTypes::DeviceInfoData>;
+    using TDeviceInfoMap = std::unordered_map<BackwardsCompatibility::DeviceIdData, FancyZonesDataTypes::DeviceInfoData>;
     using TCustomZoneSetsMap = std::unordered_map<std::wstring, FancyZonesDataTypes::CustomLayoutData>;
     using TLayoutQuickKeysMap = std::unordered_map<std::wstring, int>;
 
@@ -99,4 +127,16 @@ namespace JSONHelpers
     // replace zones-settings: custom layouts
     std::optional<TCustomZoneSetsMap> ParseCustomZoneSets(const json::JsonObject& fancyZonesDataJSON);
     void SaveCustomLayouts(const TCustomZoneSetsMap& map);
+}
+
+namespace std
+{
+    template<>
+    struct hash<BackwardsCompatibility::DeviceIdData>
+    {
+        size_t operator()(const BackwardsCompatibility::DeviceIdData& Value) const
+        {
+            return 0;
+        }
+    };
 }
